requirements.txt
asgiref==3.5.0
certifi==2021.10.8
distlib==0.3.4
dj-database-url==0.5.0
Django==3.1.14
django-ckeditor==6.2.0
django-environ==0.8.1
django-js-asset==2.0.0
django-markdown==0.8.4
django-taggit==2.1.0
filelock==3.5.1
importlib-metadata==4.11.1
Markdown==3.3.6
platformdirs==2.5.0
psycopg2-binary==2.9.3
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
virtualenv==20.13.1
virtualenv-clone==0.5.7
whitenoise==6.0.0
zipp==3.7.0
